+++ cli/cmd_group.go
@@ -53,18 +53,28 @@ func execGroup(cmd *cobra.Command, args []string) error {
 	return rc.writers.srcw.SetActiveGroup(group)
 }
 
+// newGroupListCommand is the hidden "sq lsg" command. It is
+// a convenience for "sq ls --groups --verbose".
+//
+// But, it's a real bad UX smell. It is hidden because of shame.
+// At one point, I had implemented this as "sq groups", but that
+// also didn't feel right. The "sq ls -g" command seems like the right
+// thing.
+//
+// When I have some more experience using "sq ls -gv" vs "sq lsg",
+// hopefully this will get erased from the history books.
 func newGroupListCommand() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:     "groups",
+		Use:     "lsg",
+		Hidden:  true,
 		RunE:    execGroupList,
 		Args:    cobra.NoArgs,
 		Short:   "List groups",
-		Long:    `List groups`,
-		Example: `  $ sq groups`,
+		Long:    `List groups.`,
+		Example: `  $ sq lsg`,
 	}
 
 	cmd.Flags().BoolP(flagJSON, flagJSONShort, false, flagJSONUsage)
-
 	return cmd
 }
 
@@ -72,5 +82,6 @@ func execGroupList(cmd *cobra.Command, _ []string) error {
 	rc := RunContextFrom(cmd.Context())
 	cfg := rc.Config
 
+	rc.writers.fm.Verbose = true
 	return rc.writers.srcw.Groups(cfg.Sources.ActiveGroup(), cfg.Sources.Groups())
 }